refactor: Extract video controls to separate component
