+++ application/ui/src/features/annotator/video-player/video-toolbar/video-annotator/video-timeline/video-timeline.component.tsx
@@ -22,7 +22,8 @@ const MIN_SIZE_OF_SEGMENT = 2 * 8;
 export const VideoTimeline = ({ labels }: VideoTimelineProps) => {
     const containerRef = useRef<HTMLDivElement>(null);
     const size = useSizeHook(containerRef);
-    const { videoFrame, isPlaying } = useVideoPlayer();
+    const { videoFrame, videoControls } = useVideoPlayer();
+    const { isPlaying } = videoControls;
     const frameNumber = 0;
     // TODO: fps * current time
